requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.1.0
colorama==0.4.5
commonmark==0.9.1
datasets==2.3.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dgl==0.6.1
dgllife==0.2.9
dill==0.3.5.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.7.1
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.8.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
huggingface-hub==0.8.1
hyperopt==0.2.7
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.6.0
jupyter-client==7.3.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mordred==1.2.0
multidict==6.0.2
multiprocess==0.70.13
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
notebook==6.4.12
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.1.1
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==23.2.0
rdkit-pypi==2022.3.3
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
responses==0.18.0
rich==12.4.4
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.6.0
seqeval==1.2.2
setproctitle==1.2.3
shortuuid==1.0.9
simpletransformers==0.63.7
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.3.0
streamlit==1.10.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
toolz==0.11.2
torch==1.10.2
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
wandb==0.12.19
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.0
